refactor: rename the nomatim module to nominatim
The geocoding module was spelled `nomatim`, missing the second `n`. The
service is Nominatim, and every human-readable string nearby -- including
the "Error sending request to Nominatim" text and the 429 rate-limit
message -- already spelled it correctly. Only the module identifier and
its filename carried the typo, which made it load-bearing in `use`
statements and a papercut when searching the codebase.

Mechanical: `git mv` so history follows the file, plus the three
references in main.rs and api_tests.rs. No behavioural change; the module
is internal to a binary crate, so nothing downstream can depend on it.

Closes #36
